move vm structure to its own file
