+++ src/operations/ld.rs
@@ -1,24 +1,21 @@
 use crate::{
     VMState,
     error::VMError,
-    mem_read,
     operations::utils::{sign_extend, update_flags},
     registers::Register,
 };
 pub fn handle_ld(instruction: u16, vm: &mut VMState) -> Result<(), VMError> {
     let dest_reg = ((instruction >> 9) & 0x7) as usize;
     let pc_offset = sign_extend(instruction & 0x1FF, 9);
-    vm.registers[dest_reg] = mem_read(
-        vm.registers[Register::PC.usize()].wrapping_add(pc_offset),
-        vm,
-    )?;
+    vm.registers[dest_reg] =
+        vm.mem_read(vm.registers[Register::PC.usize()].wrapping_add(pc_offset))?;
     update_flags(vm, vm.registers[dest_reg])?;
     Ok(())
 }
 
 #[cfg(test)]
 mod test {
-    use crate::{VMState, mem_write, operations::ld::handle_ld, registers::Register};
+    use crate::{VMState, operations::ld::handle_ld, registers::Register};
 
     #[test]
     fn loads_a_register() {
@@ -26,7 +23,7 @@ mod test {
         let pc_value: u16 = 0x3000;
         let memory_address = pc_value.wrapping_add(offset_u16);
         let mut vm = VMState::init().unwrap();
-        mem_write(memory_address, 50, &mut vm);
+        vm.mem_write(memory_address, 50);
         // LD   DestReg PCOffset
         // 0010 001     000000100
         let ld_ix: u16 = 0x2204;
