+++ src/operations/ldr.rs
@@ -1,21 +1,20 @@
 use crate::{
     VMState,
     error::VMError,
-    mem_read,
     operations::utils::{sign_extend, update_flags},
 };
 pub fn handle_ldr(instruction: u16, vm: &mut VMState) -> Result<(), VMError> {
     let dest_reg = ((instruction >> 9) & 0x7) as usize;
     let base_reg = ((instruction >> 6) & 0x7) as usize;
     let offset = sign_extend(instruction & 0x3F, 6);
-    vm.registers[dest_reg] = mem_read(vm.registers[base_reg].wrapping_add(offset), vm)?;
+    vm.registers[dest_reg] = vm.mem_read(vm.registers[base_reg].wrapping_add(offset))?;
     update_flags(vm, vm.registers[dest_reg])?;
     Ok(())
 }
 
 #[cfg(test)]
 mod test {
-    use crate::{VMState, mem_write, operations::ldr::handle_ldr, registers::Register};
+    use crate::{VMState, operations::ldr::handle_ldr, registers::Register};
 
     #[test]
     fn loads_register() {
@@ -24,11 +23,7 @@ mod test {
         let offset: u16 = 0x0004; //offset that will be used in ix.
         let random_memory_content: u16 = 400;
         vm.registers[Register::R2.usize()] = content_base_reg;
-        mem_write(
-            content_base_reg.wrapping_add(offset),
-            random_memory_content,
-            &mut vm,
-        );
+        vm.mem_write(content_base_reg.wrapping_add(offset), random_memory_content);
         // LDR  DestReg BaseReg Offset
         // 0110 001     010     000100
         let ldr_ix = 0x6284;
